Removed old Python type and renamed "raw" Python type to PythonFileType
