+++ src/main/resources/META-INF/services/com.atomist.rug.spi.Typed
@@ -20,9 +20,7 @@ com.atomist.rug.kind.dynamic.MutableContainerTypeProvider
 
 com.atomist.rug.kind.csharp.CSharpFileType
 
-
 com.atomist.rug.kind.python3.PythonFileType
-com.atomist.rug.kind.python3.PythonRawFileType
 
 com.atomist.rug.kind.python3.RequirementsType
 com.atomist.rug.kind.python3.PythonRequirementsTxtType
